Add a reusable Form component for Create and Edit
